+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_custom/marines2/group1/config.hpp
@@ -8,7 +8,7 @@ class 34thPRC_ArmourCustom_Marines2_Jade_Uniform_Urban_Light
 	{
 		alwaysSelectable = 1;
 		changeingame = 0;
-		values[]={"jade"};
+		values[]={"jade",};
 		class jade
 		{
 			label = "Jade";
@@ -84,7 +84,7 @@ class 34thPRC_ArmourCustom_Marines2_Jade_Uniform_Urban_Light
 	};
 };
 
-//Custom Marine Helmet Group 1: Jade
+//Custom Marine Helmet Group 1: Jade, Ramirez, Conex
 class 34thPRC_ArmourCustom_Marines2_Jade_CH252_Urban
 {
 	label = "[34thPRC] CH252 Custom Helmets";
@@ -94,10 +94,18 @@ class 34thPRC_ArmourCustom_Marines2_Jade_CH252_Urban
 	{
 		alwaysSelectable = 1;
 		changeingame = 0;
-		values[]={"jade"};
+		values[]={"jade", "ramirez", "conex"};
 		class jade
 		{
 			label = "Jade";
+		};
+		class ramirez
+		{
+			label = "Ramirez";
+		};
+		class conex
+		{
+			label = "Conex";
 		};
 	};
 	class camo