+++ src/Addons/34thPRC_ArmourCustom/data/marines2/config_weapons.hpp
@@ -1,1 +1,3 @@
 #include "jade\config_weapons.hpp"
+#include "ramirez\config_weapons.hpp"
+#include "conex\config_weapons.hpp"