+++ src/changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Updated
 - Added black hot thermals to HUL Gen 3 variants to make them leadership available
+- Yandere Recon Custom Helm
+- Ramirez Custom Marine Helm
 
 ## 0.23.0
 ### Added